kuairand-like-prediction/requirements.txt
torch>=1.12.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.60.0
matplotlib>=3.4

pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
lightgbm>=3.3
torch>=1.10
pyyaml
joblib
matplotlib
tqdm
jupyter
